+++ test/associations/has-many.test.js
@@ -117,7 +117,6 @@ describe(Support.getTestDialectTeaser("HasMany"), function() {
 
     describe('hasAll', function() {
       beforeEach(function(done) {
-        var self = this
         this.Article = this.sequelize.define('Article', {
           'title': DataTypes.STRING
         })
@@ -332,7 +331,7 @@ describe(Support.getTestDialectTeaser("HasMany"), function() {
         Article.sync({ force: true }).success(function() {
           Label.sync({ force: true }).success(function() {
             Article.create({ title: 'foo' }).success(function(article) {
-              article.createLabel({ text: 'bar' }).success(function(label) {
+              article.createLabel({ text: 'bar' }).success(function() {
 
                 Label
                   .findAll({ where: { ArticleId: article.id }})
@@ -346,8 +345,37 @@ describe(Support.getTestDialectTeaser("HasMany"), function() {
         })
       })
 
+      it('creates the object with the association directly', function(done) {
+        var spy = sinon.spy()
+
+        var Article = this.sequelize.define('Article', {
+          'title': DataTypes.STRING
+
+        }), Label   = this.sequelize.define('Label', {
+          'text': DataTypes.STRING,
+          'ArticleId': {
+            type: DataTypes.INTEGER,
+            allowNull: false
+          }
+        })
+
+        Article.hasMany(Label)
+
+        Article.sync({ force: true }).success(function() {
+          Label.sync({ force: true }).success(function() {
+            Article.create({ title: 'foo' }).success(function(article) {
+              article.createLabel({ text: 'bar' }).on('sql', spy).complete(function(err, label) {
+                expect(err).not.to.be.ok
+                expect(spy.calledOnce).to.be.true
+                expect(label.ArticleId).to.equal(article.id)
+                done()
+              })
+            })
+          })
+        })
+      })
+
       it('supports transactions', function(done) {
-        var self = this
         Support.prepareTransactionTest(this.sequelize, function(sequelize) {
           var Article = sequelize.define('Article', { 'title': DataTypes.STRING })
             , Label   = sequelize.define('Label', { 'text': DataTypes.STRING })
@@ -358,12 +386,15 @@ describe(Support.getTestDialectTeaser("HasMany"), function() {
             Label.sync({ force: true }).success(function() {
               Article.create({ title: 'foo' }).success(function(article) {
                 sequelize.transaction(function (t) {
-                  article.createLabel({ text: 'bar' }, { transaction: t }).success(function(label) {
-                    Label.findAll({ where: { ArticleId: article.id }}).success(function(labels) {
-                      expect(labels.length).to.equal(0)
-                      Label.findAll({ where: { ArticleId: article.id }}, { transaction: t }).success(function(labels) {
-                        expect(labels.length).to.equal(1)
-                        t.rollback().success(function() { done() })
+                  article.createLabel({ text: 'bar' }, { transaction: t }).success(function() {
+                    Label.findAll().success(function (labels) {
+                      expect(labels.length).to.equal(0);
+                      Label.findAll({ where: { ArticleId: article.id }}).success(function(labels) {
+                        expect(labels.length).to.equal(0)
+                        Label.findAll({ where: { ArticleId: article.id }}, { transaction: t }).success(function(labels) {
+                          expect(labels.length).to.equal(1)
+                          t.rollback().success(function() { done() })
+                        })
                       })
                     })
                   })
@@ -549,6 +580,7 @@ describe(Support.getTestDialectTeaser("HasMany"), function() {
           ])
 
           chainer.run().success(function (results, john, task1, task2) {
+            self.tasks = [task1, task2];
             john.setTasks([task1, task2]).done(function(err) {
               expect(err).not.to.be.ok
               done()
@@ -607,6 +639,46 @@ describe(Support.getTestDialectTeaser("HasMany"), function() {
         })
       })
 
+      it('supports a where not in', function (done) {
+        this.User.find({
+          where: {
+            username: 'John'
+          }
+        }).success(function (john) {
+          john.getTasks({
+            where: {
+              title: {
+                not: ['Get rich']
+              }
+            }
+          }).success(function (tasks) {
+            expect(tasks).to.have.length(1)
+            done()
+          })
+        })
+      });
+
+      it('supports a where not in on the primary key', function (done) {
+        var self = this;
+
+        this.User.find({
+          where: {
+            username: 'John'
+          }
+        }).success(function (john) {
+          john.getTasks({
+            where: {
+              id: {
+                not: [self.tasks[0].get('id')]
+              }
+            }
+          }).success(function (tasks) {
+            expect(tasks).to.have.length(1)
+            done()
+          })
+        })
+      });
+
       it("only gets objects that fulfill options with a formatted value", function(done) {
         this.User.find({where: {username: 'John'}}).success(function (john) {
           john.getTasks({where: ['active = ?', true]}).success(function (tasks) {
@@ -621,6 +693,30 @@ describe(Support.getTestDialectTeaser("HasMany"), function() {
           expect(err).not.to.be.ok
           expect(john.tasks).to.have.length(2);
           done();
+        })
+      })
+
+      it("get associated objects with an eager load with conditions but not required", function(done) {
+        var Label = this.sequelize.define('Label', { 'title': DataTypes.STRING, 'isActive': DataTypes.BOOLEAN })
+          , Task = this.Task
+          , User = this.User
+
+        Task.hasMany(Label)
+
+        Label.sync({ force: true }).success(function() {
+          User.find({
+            where: { username: 'John'},
+            include: [
+              { model: Task, required: false, include: [
+                { model: Label, required: false, where: { isActive: true } }
+              ]}
+            ]
+          }).done(function (err, john) {
+            expect(err).not.to.be.ok
+            expect(john.tasks).to.have.length(2);
+            done();
+          })
+
         })
       })
 
@@ -770,7 +866,6 @@ describe(Support.getTestDialectTeaser("HasMany"), function() {
       })
 
       it('supports transactions', function(done) {
-        var self = this
         Support.prepareTransactionTest(this.sequelize, function(sequelize) {
           var User = sequelize.define('User', { username: DataTypes.STRING })
             , Task = sequelize.define('Task', { title: DataTypes.STRING })
@@ -830,7 +925,6 @@ describe(Support.getTestDialectTeaser("HasMany"), function() {
 
     describe('optimizations using bulk create, destroy and update', function () {
       beforeEach(function (done) {
-        var self = this
         this.User = this.sequelize.define('User', { username: DataTypes.STRING }, {timestamps: false})
         this.Task = this.sequelize.define('Task', { title: DataTypes.STRING }, {timestamps: false})
 
@@ -944,7 +1038,6 @@ describe(Support.getTestDialectTeaser("HasMany"), function() {
 
     describe('primary key handling for join table', function () {
       beforeEach(function (done) {
-        var self = this
         this.User = this.sequelize.define('User',
           { username: DataTypes.STRING },
           { tableName: 'users'}
@@ -1215,7 +1308,7 @@ describe(Support.getTestDialectTeaser("HasMany"), function() {
                   expect(err).not.to.be.ok
                   Task.findAll().done(function (err, tasks) {
                     expect(err).not.to.be.ok
-                    worker.setTasks(tasks).done(function (err) {
+                    worker.setTasks(tasks).done(function () {
                       worker.setTasks(tasks).done(function (err) {
                         expect(err).not.to.be.ok
                         done()
@@ -1246,7 +1339,7 @@ describe(Support.getTestDialectTeaser("HasMany"), function() {
                 expect(err).not.to.be.ok
                 Task.findAll().done(function (err, tasks) {
                   expect(err).not.to.be.ok
-                  worker.setTasks(tasks).done(function (err) {
+                  worker.setTasks(tasks).done(function () {
                     worker.removeTask(tasks[0]).done(function (err) {
                       expect(err).not.to.be.ok
 
@@ -1364,8 +1457,7 @@ describe(Support.getTestDialectTeaser("HasMany"), function() {
       })
 
       it("correctly identifies its counterpart when through is a string", function (done) {
-        var self = this
-          , User = this.sequelize.define('User', {})
+        var User = this.sequelize.define('User', {})
           , Group = this.sequelize.define('Group', {})
 
         User.hasMany(Group, { as: 'MyGroups', through: 'group_user'})
@@ -1381,8 +1473,7 @@ describe(Support.getTestDialectTeaser("HasMany"), function() {
       })
 
       it("correctly identifies its counterpart when through is a model", function (done) {
-        var self = this
-          , User = this.sequelize.define('User', {})
+        var User = this.sequelize.define('User', {})
           , Group = this.sequelize.define('Group', {})
           , UserGroup = this.sequelize.define('GroupUser', {}, {tableName: 'user_groups'})