+++ test/mysql/query-generator.test.js
@@ -2,8 +2,6 @@
 var chai      = require('chai')
   , expect    = chai.expect
   , Support   = require(__dirname + '/../support')
-  , dialect   = Support.getTestDialect()
-  , util      = require("util")
   , _         = require('lodash')
   , QueryGenerator = require("../../lib/dialects/mysql/query-generator")
 
@@ -545,7 +543,7 @@ if (Support.dialectIsMySQL()) {
     _.each(suites, function(tests, suiteTitle) {
       describe(suiteTitle, function() {
         tests.forEach(function(test) {
-          var title = test.title || 'MySQL correctly returns ' + test.expectation + ' for ' + util.inspect(test.arguments)
+          var title = test.title || 'MySQL correctly returns ' + test.expectation + ' for ' + JSON.stringify(test.arguments)
           it(title, function(done) {
             // Options would normally be set by the query interface that instantiates the query-generator, but here we specify it explicitly
             var context = test.context || {options: {}};