+++ test/sqlite/dao-factory.test.js
@@ -124,7 +124,7 @@ if (dialect === 'sqlite') {
             })
           })
 
-          xit("should make aliased attributes available", function(done) {
+          it.skip("should make aliased attributes available", function(done) {
             this.User.find({ where: { name:'user' }, attributes: ['id', ['name', 'username']] }).success(function(user) {
               expect(user.username).to.equal('user')
               done()